Add day 16 - currently slow implementation
Use a knapsack-like approach to find the best path's score and valves
after n minutes.  This seems to work for part 1 but is kind of slow
(~1s). Fort part two, it seems to work for the real input to run that
algorithm two times: first find the best path for person 1, then find
the best path for the remaining (unopened) valves.  Unfortunately, that
approach does not work for the example input and should be refactored.